Export DType fields as public
