+++ vortex-layout/src/segments/mod.rs
@@ -52,7 +52,7 @@ pub mod test {
 
     use super::*;
     use crate::operations::Poll;
-    use crate::scanner::LayoutScan;
+    use crate::reader::LayoutReader;
     use crate::segments::SegmentReader;
     use crate::RowMask;
 
@@ -62,10 +62,10 @@ pub mod test {
     }
 
     impl TestSegments {
-        pub fn do_scan(&self, scan: Arc<dyn LayoutScan>) -> ArrayData {
+        pub fn do_scan(&self, scan: Arc<dyn LayoutReader>) -> ArrayData {
             let row_count = scan.layout().row_count();
             let mut scanner = scan
-                .create_scanner(RowMask::new_valid_between(0, row_count))
+                .create_eval(RowMask::new_valid_between(0, row_count))
                 .vortex_expect("Failed to create scanner");
             match scanner.poll(self).vortex_expect("Failed to poll scanner") {
                 Poll::Some(array) => array,
